Correção provisória da página de postagem
