+++ client/explicaaso/src/pages/PaginaEditarPost.js
@@ -1,7 +1,9 @@
-import { useState } from "react";
+import { createContext, useState } from "react";
 import RichText from "../components/EditorTexto/RichText";
 import styled from "styled-components";
-import { useCurrentEditor } from "@tiptap/react";
+import { useCurrentEditor, useEditor } from "@tiptap/react";
+import StarterKit from '@tiptap/starter-kit';
+import Underline from '@tiptap/extension-underline';
 //import { useNavigate } from "react-router-dom";
 
 const ContainerPag = styled.div`
@@ -52,9 +54,59 @@ const Botao = styled.button`
     }
 `
 
+const extensions = [
+    StarterKit.configure({
+      bulletList: {
+        keepMarks: true,
+        keepAttributes: false,
+      },
+      orderedList: {
+        keepMarks: true,
+        keepAttributes: false,
+      },
+    }),
+    Underline,
+  ]
+  
+  const content = `
+  <h2>
+    Hi there,
+  </h2>
+  <p>
+    this is a <em>basic</em> example of <strong>tiptap</strong>. Sure, there are all kind of basic text styles you’d probably expect from a text editor. But wait until you see the lists:
+  </p>
+  <ul>
+    <li>
+      That’s a bullet list with one …
+    </li>
+    <li>
+      … or two list items.
+    </li>
+  </ul>
+  <p>
+    Isn’t that great? And all of that is editable. But wait, there’s more. Let’s try a code block:
+  </p>
+  <pre><code class="language-css">body {
+  display: none;
+  }</code></pre>
+  <p>
+    I know, I know, this is impressive. It’s only the tip of the iceberg though. Give it a try and click a little bit around. Don’t forget to check the other examples too.
+  </p>
+  <blockquote>
+    Wow, that’s amazing. Good work, boy! 👏
+    <br />
+    — Mom
+  </blockquote>
+  `
+
+export const RTContext = createContext();
+
 function EditarPost(props) {
     const [titulo, setTitulo] = useState('');
-    const { editor } = useCurrentEditor();
+    const editor = useEditor({
+        extensions,
+        content,
+    });
 
     //let navigate = useNavigate();
 
@@ -64,19 +116,22 @@ function EditarPost(props) {
       //  navigate("/pagina-blog");
     }
 
-    /**ARRUMAR ESSA CACETA */
-    /**Aparentemente useContext é o que resolve essa merda, mas naõ sei usar ainda */
+    /**Esse console log demonstra a orde de renderização dos componentes */
     console.log(editor ? true : false);
 
     return(
         <ContainerPag $isMobile={props.isMobile}>
             <Titulo placeholder="Título do Post" onChange={(e) => {setTitulo(e.target.value);}}/>
-            <RichText isMobile={props.isMobile}/>
+            
+            <RTContext.Provider value={editor}>
+                <RichText isMobile={props.isMobile} />
+            </RTContext.Provider>
+            
             <ContainerBotao>
                 <Botao $isMobile={props.isMobile} onClick={criarPost}>Postar</Botao>
             </ContainerBotao>
             {/* <pre>
-            { editor ? editor.getJSON() : null }
+            { editor != null && editor.getHTML() }
             </pre> */}
         </ContainerPag>
     );
